dom/media/test/test_texttrack_moz.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for Bug 881976 - TextTrackCue Computed Position</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<video id="v" src="seek.webm" preload="metadata">
<script type="text/javascript">

function runTest() {
  var video = document.getElementById("v");

  // Check if adding a text track manually sets the TextTrackList correctly.

  // HTMLTrackElement.textTrackList is an extension available only to
  // privileged code, so we need to access it through the SpecialPowers
  // object.
  video.addTextTrack("subtitles", "", "");
  is(SpecialPowers.unwrap(SpecialPowers.wrap(video.textTracks[0]).textTrackList),
      video.textTracks,
      "The Track's TextTrackList should be the Video's TextTrackList.");

  var trackElement = document.createElement("track");
  trackElement.src = "basic.vtt";
  trackElement.kind = "subtitles";

  video.appendChild(trackElement);
  video.addEventListener("loadedmetadata", function run_tests() {
    // Re-que run_tests() at the end of the event loop until the track
    // element has loaded its data.
    if (trackElement.readyState == HTMLTrackElement.LOADING) {
      setTimeout(run_tests, 0);
      return;
    }
    is(trackElement.readyState, HTMLTrackElement.LOADED,
        "Track::ReadyState should be set to LOADED.");
    is(SpecialPowers.unwrap(SpecialPowers.wrap(trackElement.track).textTrackList),
       video.textTracks,
       "TrackElement's Track's TextTrackList should be the Video's TextTrackList.");

    SimpleTest.finish();
  });
}

SimpleTest.waitForExplicitFinish();
onload = runTest;

</script>
</body>
</html>
